Sidebar box to navigate through learning path
...When viewing an item in that learning path
