+++ app/helpers/learning_paths_helper.rb
@@ -1,0 +1,7 @@
+module LearningPathsHelper
+
+  def learning_path_breadcrumb_param(topic_link, topic_item)
+    { lp: [topic_link, topic_item].map(&:id).join(':') }
+  end
+
+end
