+++ app/views/events/_event.html.erb
@@ -1,6 +1,8 @@
+<% learning_path_topic_item ||= nil %>
+<% collection_item ||= learning_path_topic_item %>
 <li class="masonry-brick media-item long">
   <%= link_to event, class: 'link-overlay with-left-icon' do %>
-    <%= item_order_badge(collection_item) if defined? collection_item %>
+    <%= item_order_badge(collection_item) if collection_item %>
     <div class="with-left-icon">
       <div class="icon-container hidden-xs">
         <i class="icon icon-h2 <%= event.presence %>-event-icon"></i><br/>
@@ -52,6 +54,6 @@
       <% end %>
     </div>
 
-    <%= item_comment(collection_item) if defined? collection_item %>
+    <%= item_comment(collection_item) if collection_item %>
   <% end %>
 </li>
